+++ src/main/java/webserver/controller/UserCreateController.java
@@ -12,7 +12,7 @@ import http.HttpStatus;
 import http.request.HttpRequest;
 import http.response.HttpResponse;
 import model.User;
-import webserver.persist.Users;
+import webserver.persistence.Users;
 
 public class UserCreateController implements Controller {
